Moves level set calculation to multibody/hydroelastics
