+++ multibody/hydroelastics/contact_surface_calculator.h
